+++ .ansible/collections/ansible_collections/netapp/ontap/plugins/module_utils/netapp.py
@@ -49,7 +49,7 @@ try:
 except ImportError:
     ANSIBLE_VERSION = 'unknown'
 
-COLLECTION_VERSION = "23.3.0"
+COLLECTION_VERSION = "23.4.0"
 CLIENT_APP_VERSION = "%s/%s" % ("%s", COLLECTION_VERSION)
 IMPORT_EXCEPTION = None