+++ tests/unit/plugins/modules/test_na_ontap_vserver_audit.py
@@ -1,4 +1,4 @@
-# (c) 2023, NetApp, Inc
+# (c) 2023-2026, NetApp, Inc
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 ''' unit tests for Ansible module: na_ontap_vserver_audit '''
@@ -8,6 +8,7 @@ from __future__ import (absolute_import, division, print_function)
 __metaclass__ = type
 
 import pytest
+import sys
 
 from ansible_collections.netapp.ontap.tests.unit.compat.mock import patch
 import ansible_collections.netapp.ontap.plugins.module_utils.netapp as netapp_utils
@@ -19,8 +20,9 @@ from ansible_collections.netapp.ontap.tests.unit.framework.rest_factory import r
 from ansible_collections.netapp.ontap.plugins.modules.na_ontap_vserver_audit \
     import NetAppONTAPVserverAudit as my_module, main as my_main      # module under test
 
-if not netapp_utils.has_netapp_lib():
-    pytestmark = pytest.mark.skip('skipping as missing required netapp_lib')
+if not netapp_utils.HAS_REQUESTS and sys.version_info < (2, 7):
+    pytestmark = pytest.mark.skip(
+        'Skipping Unit Tests on 2.6 as requests is not available')
 
 # REST API canned responses when mocking send_request
 SRR = rest_responses({